backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
asyncpg==0.29.0
pgvector==0.2.4
openai==1.12.0
pydantic==2.5.3
python-dotenv==1.0.1
alembic==1.13.1
httpx==0.26.0
numpy>=1.26.0
slowapi==0.1.9
